style: fix all clippy warnings with --all-features --all-targets
- Remove unnecessary drop() calls on non-Drop types (parser/scan.rs, progressive.rs)
- Fix manual assign operations (idct_int.rs, arithmetic.rs)
- Fix unnecessary casts (idct_int.rs, huffman/encode.rs)
- Fix collapsible if statements (parser/output.rs, progressive.rs)
- Fix unnecessary reference creation (scanline.rs, examples)
- Remove unused imports (examples, tests)
- Fix doc comment placement for ultrahdr_reader (was split across functions)
- Allow while_immutable_condition for FFI loops where condition is mutated by C
- Fix abs().to_u8() pattern (decode_accuracy_corpus.rs)
- Fix manual RangeInclusive::contains (arithmetic_tolerance.rs)

+++ zenjpeg/src/huffman/encode.rs
@@ -509,7 +509,7 @@ impl HuffmanDecodeTable {
         }
         if code_length <= 16 {
             let symbol_bits = bits16 >> (16 - code_length);
-            let idx = (symbol_bits as i32 + self.valoffset[code_length]) as usize;
+            let idx = (symbol_bits + self.valoffset[code_length]) as usize;
             if idx < self.values.len() {
                 return Some((self.values[idx], code_length as u8));
             }
